requirements.txt
certifi==2023.7.22
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
duckdb==0.6.1
fonttools==4.43.0
greenlet==2.0.1
idna==3.7
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
mysql-connector-python
numpy==1.24.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==22.0
pandas==1.5.2
pandasql==0.7.3
Pillow==10.3.0
presto-python-client==0.8.3
prettytable==3.5.0
protobuf==3.20.2
psutil==5.9.5
psycopg2-binary==2.9.5
py4j==0.10.9.5
pyparsing==3.0.9
pyspark==3.3.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.32.0
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
SQLAlchemy==1.4.45
threadpoolctl==3.1.0
torch==1.13.1
typing_extensions==4.4.0
urllib3==1.26.18
wcwidth==0.2.5
